Убран неиспользуемый код, show-popup и show-alert перенесены в папку utils
